requirements.txt
fastapi
uvicorn
langchain>=0.1.20
langchain-openai
langchain-anthropic
langchain-cohere
langchain-google-genai
openai
python-dotenv
python-multipart
pydantic
itsdangerous
pypdf
serpapi
firebase-admin
google-auth
requests
